+++ app/resources/views/pages/dominion/magic.blade.php
@@ -40,7 +40,7 @@
                                                 <label for="target_dominion">Select a target</label>
                                                 <select name="target_dominion" id="target_dominion" class="form-control select2" required style="width: 100%" data-placeholder="Select a target dominion" {{ $selectedDominion->isLocked() ? 'disabled' : null }}>
                                                     <option></option>
-                                                    @foreach ($rangeCalculator->getDominionsInRange($selectedDominion, true) as $dominion)
+                                                    @foreach ($rangeCalculator->getDominionsInRange($selectedDominion, true, true) as $dominion)
                                                         <option value="{{ $dominion->id }}"
                                                                 data-race="{{ $dominion->race->name }}"
                                                                 data-land="{{ number_format($landCalculator->getTotalLand($dominion)) }}"
@@ -48,6 +48,7 @@
                                                                 data-war="{{ $governmentService->isAtWar($selectedDominion->realm, $dominion->realm) ? 1 : 0 }}"
                                                                 data-revenge="{{ in_array($dominion->id, $recentlyInvadedByDominionIds) ? 1 : 0 }}"
                                                                 data-guard="{{ $guardMembershipService->isBlackGuardMember($dominion) && $guardMembershipService->isBlackGuardMember($selectedDominion) ? 1 : 0 }}"
+                                                                data-friendly="{{ $selectedDominion->realm_id == $dominion->realm_id }}"
                                                             >
                                                             {{ $dominion->name }} (#{{ $dominion->realm->number }})
                                                         </option>
@@ -141,6 +142,47 @@
                                                                 name="spell"
                                                                 value="{{ $spell->key }}"
                                                                 class="btn btn-primary btn-block war-spell disabled"
+                                                                {{ $selectedDominion->isLocked() || $selectedDominion->round->hasOffensiveActionsDisabled() || !$canCast || (now()->diffInDays($selectedDominion->round->start_date) < 3) ? 'disabled' : null }}>
+                                                            {{ $spell->name }}
+                                                        </button>
+                                                        <p style="margin: 5px 0;">{{ $spellHelper->getSpellDescription($spell) }}</p>
+                                                        <small>
+                                                            @if ($canCast)
+                                                                Mana cost: <span class="text-success">{{ number_format($spellCalculator->getManaCost($selectedDominion, $spell)) }}</span><br/>
+                                                            @else
+                                                                Mana cost: <span class="text-danger">{{ number_format($spellCalculator->getManaCost($selectedDominion, $spell)) }}</span><br/>
+                                                            @endif
+                                                            @if ($spell->duration)
+                                                                Lasts {{ $spell->duration }} hours<br/>
+                                                            @endif
+                                                            @if (!empty($spell->races))
+                                                                Racial<br/>
+                                                            @endif
+                                                        </small>
+                                                    </div>
+                                                </div>
+                                            @endforeach
+                                        </div>
+                                    @endforeach
+
+                                    <div class="row">
+                                        <div class="col-md-12">
+                                            <label>Friendly Spells</label>
+                                        </div>
+                                    </div>
+
+                                    @foreach ($spellHelper->getSpells($selectedDominion->race, 'friendly')->chunk(4) as $spells)
+                                        <div class="row">
+                                            @foreach ($spells as $spell)
+                                                @php
+                                                    $canCast = $spellCalculator->canCast($selectedDominion, $spell);
+                                                @endphp
+                                                <div class="col-xs-6 col-sm-3 col-md-6 col-lg-3 text-center">
+                                                    <div class="form-group">
+                                                        <button type="submit"
+                                                                name="spell"
+                                                                value="{{ $spell->key }}"
+                                                                class="btn btn-primary btn-block friendly-spell disabled"
                                                                 {{ $selectedDominion->isLocked() || $selectedDominion->round->hasOffensiveActionsDisabled() || !$canCast || (now()->diffInDays($selectedDominion->round->start_date) < 3) ? 'disabled' : null }}>
                                                             {{ $spell->name }}
                                                         </button>
@@ -267,10 +309,16 @@
                 var warStatus = $(this).find(":selected").data('war');
                 var revengeStatus = $(this).find(":selected").data('revenge');
                 var guardStatus = $(this).find(":selected").data('guard');
-                if (warStatus == 1 || revengeStatus == 1 || guardStatus == 1) {
+                var friendlyStatus = $(this).find(":selected").data('friendly');
+                if (!friendlyStatus && (warStatus == 1 || revengeStatus == 1 || guardStatus == 1)) {
                     $('.war-spell').removeClass('disabled');
                 } else {
                     $('.war-spell').addClass('disabled');
+                }
+                if (friendlyStatus) {
+                    $('.friendly-spell').removeClass('disabled');
+                } else {
+                    $('.friendly-spell').addClass('disabled');
                 }
             });
             @if ($targetDominion)
@@ -292,6 +340,7 @@
             const war = state.element.dataset.war;
             const revenge = state.element.dataset.revenge;
             const guard = state.element.dataset.guard;
+            const friendly = state.element.dataset.friendly;
             let difficultyClass;
 
             if (percentage >= 133) {
@@ -305,7 +354,9 @@
             }
 
             warStatus = '';
-            if (war == 1) {
+            if (friendly == 1) {
+                warStatus = '<div class="pull-left">&nbsp;|&nbsp;<span class="text-green">FRIENDLY</span></div>';
+            } else if (war == 1) {
                 warStatus = '<div class="pull-left">&nbsp;|&nbsp;<span class="text-red">WAR</span></div>';
             } else if (guard == 1) {
                 warStatus = '<div class="pull-left">&nbsp;|&nbsp;<span class="text-red">SHADOW LEAGUE</span></div>';